+++ packages/iov-core/src/index.ts
@@ -7,7 +7,7 @@ export { Address, Nonce, SendTx, SetNameTx, TokenTicker, TransactionKind } from 
 export { bnsConnector, bnsFromOrToTag, bnsNonceTag, bnsSwapQueryTags } from "@iov/bns";
 export {
   Ed25519HdWallet,
-  Ed25519KeyringEntry,
+  Ed25519Wallet,
   HdPaths,
   Keyring,
   UserProfile,