+++ packages/iov-core/types/index.d.ts
@@ -1,5 +1,5 @@
 export { Address, Nonce, SendTx, SetNameTx, TokenTicker, TransactionKind } from "@iov/bcp-types";
 export { bnsConnector, bnsFromOrToTag, bnsNonceTag, bnsSwapQueryTags } from "@iov/bns";
-export { Ed25519HdWallet, Ed25519KeyringEntry, HdPaths, Keyring, UserProfile, Wallet, WalletId, WalletImplementationIdString, WalletSerializationString, } from "@iov/keycontrol";
+export { Ed25519HdWallet, Ed25519Wallet, HdPaths, Keyring, UserProfile, Wallet, WalletId, WalletImplementationIdString, WalletSerializationString, } from "@iov/keycontrol";
 export { ChainId } from "@iov/tendermint-types";
 export { MultiChainSigner } from "./multichainsigner";