+++ packages/iov-keycontrol/src/keyring.ts
@@ -5,7 +5,7 @@ import { Slip10RawIndex } from "@iov/crypto";
 import { ValueAndUpdates } from "@iov/stream";
 import { ChainId, PublicKeyBundle, SignatureBytes } from "@iov/tendermint-types";
 
-import { Ed25519HdWallet, Ed25519KeyringEntry, Secp256k1HdWallet } from "./keyring-entries";
+import { Ed25519HdWallet, Ed25519Wallet, Secp256k1HdWallet } from "./keyring-entries";
 
 export type KeyringSerializationString = string & As<"keyring-serialization">;
 export type WalletImplementationIdString = string & As<"wallet-implementation-id">;
@@ -56,7 +56,7 @@ export class Keyring {
   }
 
   private static readonly deserializationRegistry = new Map([
-    ["ed25519", (data: WalletSerializationString) => new Ed25519KeyringEntry(data)],
+    ["ed25519", (data: WalletSerializationString) => new Ed25519Wallet(data)],
     ["ed25519-hd", (data: WalletSerializationString) => new Ed25519HdWallet(data)],
     ["secp256k1-hd", (data: WalletSerializationString) => new Secp256k1HdWallet(data)],
   ] as ReadonlyArray<[string, KeyringEntryDeserializer]>);
@@ -148,7 +148,7 @@ export interface Wallet {
 
   // createIdentity will create one new identity
   readonly createIdentity: (
-    options: Ed25519KeyringEntry | ReadonlyArray<Slip10RawIndex> | number,
+    options: Ed25519Wallet | ReadonlyArray<Slip10RawIndex> | number,
   ) => Promise<LocalIdentity>;
 
   // Sets a local label associated with the public identity to be displayed in the UI.