+++ packages/iov-keycontrol/src/userprofile.ts
@@ -25,7 +25,7 @@ import {
   Wallet,
   WalletId,
 } from "./keyring";
-import { Ed25519KeyringEntry } from "./keyring-entries";
+import { Ed25519Wallet } from "./keyring-entries";
 import { DatabaseUtils } from "./utils";
 
 const { toAscii, fromBase64, toBase64, fromUtf8, toUtf8, toRfc3339, fromRfc3339 } = Encoding;
@@ -183,7 +183,7 @@ export class UserProfile {
   // creates an identitiy in the n-th keyring entry of the primary keyring
   public async createIdentity(
     id: WalletId,
-    options: Ed25519KeyringEntry | ReadonlyArray<Slip10RawIndex> | number,
+    options: Ed25519Wallet | ReadonlyArray<Slip10RawIndex> | number,
   ): Promise<LocalIdentity> {
     const entry = this.entryInPrimaryKeyring(id);
     return entry.createIdentity(options);